Begin the project.
I recently had the opportunity to introduce my wife to a board game that
she had never played before, but that I remembered from my childhood.

After an hour of setup and half an hour of playing, her only words were:

> This might be more interesting if a computer was doing all the work.

We did not finish the game.

We'll see whether or not I finish this project, and whether my wife's
prediction will be correct.