Add Park-Miller psuedo-random number generator.
Generating pseudo-random numbers is a critical function for a luck-based
strategy game simulator. I decided to take an off-the-shelf solution to
start with, but I intentionally looked for a short-and-simple function
that I could read, understand, and hopefully learn from.

I started this evening with a Google search:

> pseudorandom number generator javascript

I knew that I wanted a function that would accept a seed, so the second
search result immediately jumped out at me:

> A very simple, seedable JavaScript PRNG. - GitHub
> https://gist.github.com/blixt/f17b47c62508be59987b

I fell down a bit of a rabbit trail with the linked mathematical essays:
http://www.firstpr.com.au/dsp/rand31/
http://www.math.sci.hiroshima-u.ac.jp/%7Em-mat/MT/emt.html

...but I managed to get back to coding in less than an evening. Success.