Add independent dice, defending player wins ties.
One potential outcome for this project would be that I would
programmatically arrive at a mathematically optimal algorithm for
potential future board games with my wife.

In pursuit of this goal, I am attempting to isolate conditions that are
caused by chance from those caused by choice.

In other words, given a set of predetermined die rolls, I want to be
able to compare all the available decisions I could have made to
optimize my play. Of course, in a real (fair) game I can't know what my
next roll will be. But in this simulation, I can test *every* potential
choice against the same predetermined psuedo-random die rolls.

There is one decision that feels important to me, but I haven't been
able to fully finalize in my head yet. Is it correct to use
independently-seeded psuedo-random number generators for each player?

I think I may be falling into a bit of a fallacy here with a concern the
next roll that "belongs" to a certain player may end up going to the
other player based on choices that are made during a game. In theory, it
shouldn't matter at all as the numbers are "random" in the sense that
actions made during a turn don't *actually* affect the randomness of the
dice even though with a PNRG, technically they do.